+++ src/lib/knobs/Color.svelte
@@ -27,6 +27,7 @@
 
 	input {
 		width: 100%;
+		height: 100%;
 	}
 
 	.picker {