+++ src/lib/knobs/Number.svelte
@@ -13,3 +13,10 @@
 	<span>{name}</span>
 	<input type="number" bind:value {step}>
 </label>
+
+<style>
+	input {
+		width: 100%;
+		height: 100%;
+	}
+</style>