+++ src/lib/knobs/String.svelte
@@ -11,3 +11,9 @@
 	<input bind:value>
 </label>
 
+<style>
+	input {
+		width: 100%;
+		height: 100%;
+	}
+</style>