+++ CHANGELOG.md
@@ -2,9 +2,10 @@
 
 ## 0.1.12 (unreleased)
 
-- Change keyframes data structure
-- Make keyframe tracks toggleable
-- Color coded keyframe tracks
+- [Keyframe plugin] Change keyframes data structure
+- [Keyframe plugin] Make keyframe tracks toggleable
+- [Keyframe plugin] Color coded keyframe tracks
+- [Keyframe plugin] Number editor for selected point
 
 ## 0.1.11